Refoctored code logic into classes
